refactor: break config circular import
